string_util::format() expects an int for %d, change atof() to atoi()
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@4493 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/CmdLineOptions.cxx
@@ -915,7 +915,7 @@ void parse(int argc, char **argv, CmdLineOptions &options)
 	fprintf(stderr, "argument expected for \"%s\"\n", argv[i]);
 	usage(argv[0]);
       }
-      BZDB->set(StateDatabase::BZDB_WORLDSIZE, string_util::format("%d",atof(argv[i])*2));
+      BZDB->set(StateDatabase::BZDB_WORLDSIZE, string_util::format("%d",atoi(argv[i])*2));
       fprintf(stderr, "using world size of \"%f\"\n", BZDB->eval(StateDatabase::BZDB_WORLDSIZE));
     } else {
       fprintf(stderr, "bad argument \"%s\"\n", argv[i]);
